Refactor test commands
