+++ packages/cudos-cmd/cudos-cmd.js
@@ -1,11 +1,7 @@
 #!/usr/bin/env node
 
-const {
-  hideBin
-} = require('yargs/helpers')
-const {
-  commands
-} = require('./commands')
+const { hideBin } = require('yargs/helpers')
+const { commands } = require('./commands')
 
 async function main() {
   const args = hideBin(process.argv)
